feat: register native BAGEL ComfyUI nodes
