+++ __init__.py
@@ -1,8 +1,5 @@
-"""
-ComfyUI-Bagel - ComfyUI custom node package for the BAGEL multimodal model
-"""
+"""ComfyUI-BAGEL custom node package."""
 
 from .nodes import NODE_CLASS_MAPPINGS, NODE_DISPLAY_NAME_MAPPINGS
 
-# Export node mappings for ComfyUI
 __all__ = ["NODE_CLASS_MAPPINGS", "NODE_DISPLAY_NAME_MAPPINGS"]
